+++ semana10/labex/src/App/App.js
@@ -1,4 +1,4 @@
-import AppContainer from '../Components/AppContainer/AppContainer';
+import AppContainer from '../components/AppContainer/AppContainer';
 
 export default function App() {
   return (
